requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai>=2.5.0
pillow==10.1.0
web3==6.12.0
werkzeug==3.0.1
uuid==1.30
requests==2.31.0
google-cloud-storage==2.10.0
flask-socketio==5.3.6